prulong/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.11.0
argcomplete==3.6.2
arrow==1.3.0
attrs==25.3.0
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
backoff==2.2.1
bcrypt==4.3.0
block-sparse-attn==0.0.1
boto3==1.40.37
botocore==1.40.37
brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
circuitbreaker==2.1.3
click==8.3.0
contourpy==1.3.3
cramjam==2.11.0
cryptography==44.0.3
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.19.1
flash-attn==2.6.3
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gql==4.0.0
graphql-core==3.2.6
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
isodate==0.7.2
jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
modelscope==1.30.0
mosaicml-cli==0.5.34
mosaicml-streaming==0.8.1
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oci==2.160.2
packaging==25.0
pandas==2.3.2
paramiko==3.5.1
pillow==11.3.0
platformdirs==4.4.0
prompt-toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.6.0
pyopenssl==24.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytz==2025.2
pyyaml==6.0.2
questionary==2.1.1
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rsa==4.9.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
s3transfer==0.14.0
safetensors==0.6.2
sentencepiece==0.1.97
sentry-sdk==2.38.0
setproctitle==1.3.7
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.7.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.53.3
triton==3.4.0
types-python-dateutil==2.9.0.20250822
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
wandb==0.17.3
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.22.0
zstd==1.5.5.1
